[Update] From Branch Clario (#10)
* [Add] publid-build

* [Add] views, and init for controller and models

* [Update} Login and Register Feature

* [Update] Integrate Account Page

* [Add] view details, added GameSeeder

* [Update] Integrate On Progress BackEnd and FrontEnd

* [Update] From Clario

* [Update] Input Validation from all User

* [Delete] Clear public image file

+++ rental_aja/app/Http/Controllers/GameController.php
@@ -12,9 +12,13 @@ use App\Models\GamePlatform;
 use App\Models\Platform;
 use App\Models\GameImage;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class GameController extends Controller
 {
+
+    // ========================================= USER =========================================
+    // ------------------------------ SHOW ALL GAMES (DASHBOARD) ------------------------------
     /**
      * Display a listing of the resource.
      *
@@ -22,9 +26,33 @@ class GameController extends Controller
      */
     public function index()
     {
-        $games = Game::with(['publisher', 'genres', 'gameImages', 'platforms'])->get();
+        // get all
+        // $games = Game::with(['publisher', 'genres', 'gameImages', 'platforms'])->get();
+
+        $games = Game::with(['gameImages'])->get(['id', 'title', 'price']);
+
         return view('games.index', ['games' => $games]);
+
+        // ini masih bisa nembak buat ambil data - data yang seharusnya ga boleh didump spt publisher_id
     }
+
+    // ------------------------------ SHOW SPECIFIC GAME DETAIL ------------------------------
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        $game = Game::find($id)->load(['publisher', 'genres', 'gameImages', 'platforms']);
+        return view('games.show', ['game' => $game]);
+    }
+
+
+
+    // ========================================= ADMIN =========================================
+    // --------------------------------- SHOW ADD GAME PAGE  -----------------------------------
 
     /**
      * Show the form for creating a new resource.
@@ -33,6 +61,10 @@ class GameController extends Controller
      */
     public function create()
     {
+        if (!auth()->user() || auth()->user()->role !== 'admin') {
+            return abort(403);
+        }
+
         $publishers = Publisher::all();
         $genres = Genre::all();
         $platforms = Platform::all();
@@ -43,6 +75,8 @@ class GameController extends Controller
         ]);
     }
 
+
+    // --------------------------------- POST REQUEST ADD GAME  -----------------------------------
     /**
      * Store a newly created resource in storage.
      *
@@ -51,6 +85,23 @@ class GameController extends Controller
      */
     public function store(Request $request)
     {
+        if (!auth()->user() || auth()->user()->role !== 'admin') {
+            return abort(403);
+        }
+
+        $request->validate([
+            'title' => 'required|string|max:255',
+            'release_date' => 'required|date',
+            'description' => 'required|string|max:4096',
+            'rating' => 'required|string|in:E,E10,T,M,A',
+            'price' => 'required|numeric|between:0,9999.99',
+            'publisher_id' => 'required|integer|exists:publishers,id',
+            'genre_id' => 'required|array|min:1|exists:genres,id',
+            'platform_id' => 'required|array|min:1|exists:platforms,id',
+            'images' => 'required|array|min:1|max:5',
+            'images.*' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+        ]);
+
         $data = $request->except(['genre_id', 'platform_id', 'images']);
         $genres = $request->genre_id;
         $platforms = $request->platform_id;
@@ -73,7 +124,7 @@ class GameController extends Controller
         }
 
         foreach ($images as $image) {
-            $filename = Storage::putFile('images', $image);
+            $filename = Storage::putFile('images/game-images', $image);
             GameImage::create([
                 'path' => $filename,
                 'game_id' => $game->id
@@ -83,18 +134,8 @@ class GameController extends Controller
         return redirect('/games');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        $game = Game::find($id)->load(['publisher', 'genres', 'gameImages', 'platforms']);
-        return view('games.show', ['game' => $game]);
-    }
 
+    // --------------------------------- SHOW EDIT GAME PAGE -----------------------------------
     /**
      * Show the form for editing the specified resource.
      *
@@ -103,6 +144,10 @@ class GameController extends Controller
      */
     public function edit($id)
     {
+        if (!auth()->user() || auth()->user()->role !== 'admin') {
+            return abort(403);
+        }
+
         $publishers = Publisher::all();
         $genres = Genre::all();
         $platforms = Platform::all();
@@ -115,6 +160,8 @@ class GameController extends Controller
         ]);
     }
 
+
+    // --------------------------------- POST REQUEST EDIT GAME  -----------------------------------
     /**
      * Update the specified resource in storage.
      *
@@ -124,8 +171,23 @@ class GameController extends Controller
      */
     public function update(Request $request, Game $game)
     {
-        // $game = Game::find($id)->load(['publisher', 'genres', 'gameImages', 'platforms']);
-        // dd($game);
+        if (!auth()->user() || auth()->user()->role !== 'admin') {
+            return abort(403);
+        }
+
+        $request->validate([
+            'title' => 'required|string|max:255',
+            'release_date' => 'required|date',
+            'description' => 'required|string|max:4096',
+            'rating' => 'required|string|in:E,E10,T,M,A',
+            'price' => 'required|numeric|between:0,9999.99',
+            'publisher_id' => 'required|integer|exists:publishers,id',
+            'genre_id' => 'required|array|min:1|exists:genres,id',
+            'platform_id' => 'required|array|min:1|exists:platforms,id',
+            'images' => 'required|array|min:1|max:5',
+            'images.*' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+        ]);
+
         $data = $request->except(['genre_id', 'platform_id', 'images']);
         $genres = $request->genre_id;
         $platforms = $request->platform_id;
@@ -140,7 +202,7 @@ class GameController extends Controller
         GamePlatform::where('game_id', $game->id)->delete();
 
         foreach ($images as $image) {
-            $filename = Storage::putFile('images', $image);
+            $filename = Storage::putFile('images/game-images', $image);
             GameImage::create([
                 'path' => $filename,
                 'game_id' => $game->id
@@ -166,6 +228,7 @@ class GameController extends Controller
         return redirect('/games/' . $game->id);
     }
 
+    // --------------------------------- DELETE METHOD REQUEST DELETE GAME  -----------------------------------
     /**
      * Remove the specified resource from storage.
      *
@@ -174,6 +237,10 @@ class GameController extends Controller
      */
     public function destroy(Game $game)
     {
+        if (!auth()->user() || auth()->user()->role !== 'admin') {
+            return abort(403);
+        }
+
         foreach ($game->gameImages as $image) {
             Storage::delete($image->path);
         }
